Move some event listener class from diagram package to event_listener package
