+++ src/main/java/jp/ex_t/kazuaki/change_vision/event_listener/ClassDiagramEventListener.kt
@@ -1,12 +1,12 @@
 /*
- * ClassDiagram.kt - pair-modeling-prototype
+ * ClassDiagramEventListener.kt - pair-modeling-prototype
  * Copyright © 2021 HyodaKazuaki.
  *
  * Released under the MIT License.
  * see https://opensource.org/licenses/MIT
  */
 
-package jp.ex_t.kazuaki.change_vision.diagram
+package jp.ex_t.kazuaki.change_vision.event_listener
 
 import com.change_vision.jude.api.inf.model.*
 import com.change_vision.jude.api.inf.presentation.ILinkPresentation
